Rewrite file state insertion queries to non failing ones
This mitigates the problem when the state is inserted for the trasfer which is already deleted from the DB.
It happens because the insertion is done in the event handler, meaning it's delayed a bit.
The queries used a subquery SELECT which upon returning no rows is interpreted by INSERT statement as a single row with NULL values, causing foreign key inssue.
The rewritten ones use the INSERT SELECT statement which is free from this problem.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>