Prevent deadlocks in testsuite
When doing lock.acquire() there must be a symmetric
lock.release() call, however if the exception is thrown
in between, ten the release is not issued and there's a deadlock.

To combat this, there's `with lock` trick in Python which creates
a context which then gets collected at the very end. Similar to RAII

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>