Allow constructing outgoing files without the access to the file system.
This allows for a fairly easy reconstruction of transfer from the database upon instance startup.
Previously when restoring a file failed the whole transfer was ignored since it wasn't possible to restore its data structure.
With these changes, the structure is restored and the file system is only consulted during the regular operation of the connection handler.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>