Refactor Makefile and introduce coverage
- Makefile was refactored so now test directory
is the working directory. This cleans up
the paths as else Dockerized scripts expect that they are in
libdrop's root while the Makefile expects that it's in test dir.
Makefile also allows for easy testing and coverage. Assuming we're
in libdrop's root:
  * to test everything: make -C test
  * to test one scenario: $SCENARIO=scenario1 make -C test
  * to cover everything: make -C test coverage
  * to cover one scenario: $SCENARIO=scenario1 make -C test coverage

- Coverage requires building instrumented build of the library
and passing `LLVM_PROFILE_FILE` to produce the profile file.

This still creates default profraw files when building the library
as build.rs doesn't seem to listen to the `LLVM_PROFILE_FILE`.

For coverage a nightly version of `rustc` is required,
`rustc 1.68.0-nightly (574b64a97 2022-12-31)` is a good version.

Once coverage is built it's converted to HTML and presented via
HTTP service for easy manual inspection. Possible improvements
would be to upload lconv file somewhere for storage.

Note that all coverage files are merged automatically so cleanup
might be required before checking individual coverages.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>