Do not store the remote state, instead sync all file states upon connecting
We decided that it may be simpler to not make a diff of local and remote states upon connecting to find out what needs to be communicated, but instead send all local states to the peer.
This frees us from keeping track of what peer knows about file states since we allways sync the full picture of a transfer.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>