Restore transfer state before spawning server
Transfer state was resumed and outgoing transfer
sync was initiated before the server start.

This resulted in a lot of errors thrown around
if the server start failed.

In order to properly solve the problem, the resume step
is split into two parts: restore state and spawn outgoing
transfer sync tasks.

Without split there's a time window where we spawn the server
and the peer  syncs transfers with us. Because the state is not there
the sync would fail.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>