Reorganize receiving file stream. Use a single channel for all messages except `Progress`
This commit is an attempt to fix the problem when the sender receives a transfer cancelation before the file download is done message.
It happens because the cancelation request uses a separate channel that collects the app requests than the download task, which uses its own channel.
Thus, message reordering can happen.
This change implements sending the critical file state changes messages down the very same channel for the app requests,
which should guarantee in order processing.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>