Move testfiles into Docker image
Using `dd` proved to cause delays when running
tests in parallel. It becomes so slow that the system
becomes unresponsive and it's hard to detect that
as CPU and MEM stats look just fine.

Testfiles are basically static so we also
gain a slight benefit that this reduces memory
consumption as files are COW meaning only one
copy will persist across all scenarios at once
until a write happens(majority of testcases are read only).

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>