Parallelize testrunner
In order to paralellize testcases we can't
hardcode peer IPs anymore thus we rely on
Docker DNS to resolve peers based on name.

Peer names are constructed as `PEER-scenario`.
All containers for each scenario are spinned
at once and `peer_resolver.py` aids in resolving
the IPs at runtime and also reverse lookup
from IP back to the peer name.

runner.py has these flags:
--testcase-timeout=$(TESTCASE_TIMEOUT)
--scenarios-at-once=$(SCENARIOS_AT_ONCE)

If the peer goes down or is not present at all
Docker DNS doesn't have peer name and will
fail to resolve it, thus:
  - we must have peer present in the scenario
    doing nothing if we refer to it.
  - peer must be online for the duration
    we're refering to it.

Issues to be fixed:
* Spinning docker containers even with
`Detach=True` still takes a second or two,
so each testcase must be aware that peer
might not be available for a short while

* scenario28: IPv6 testcase fails.
Both ipv4 only and ipv6 networks are
created and I can inspect manually that
peers can telnet into each other but
something is missing to make it correct.

* scenario19-2: invalid chars fails.
This testcase changes unsupported symbols
into underscores and due to moving file
generation from Python to bash script
a small difference is introduced somewhere
and the file can't be found. It is present
in the container though when doing manual
inspection.

* Some tests are more flaky from my observations
thus this commit introduces some additional delays
and some more are most probably needed.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>