Send `TransferRequest` message down the WS only after checking if the transfer is alive.
It is a part of `TransferRequest` and `TransferCancelled` event pair suppersion.
Previously the check for transfer being alive happened before the connection was established.
In case when the transfer was cancelled while establishing the connection, it still slipped trough.
The root cause was that the WS message was send before registering the transfer as connection in the manager.
This change filps the order of operations, the transfer is registered, and then only if it's succesful the message is sent.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>