Implement norddrop_set_peer_state for tests
Implement the API call in testrunner
and the support code

Modify testcases to follow the new API
as transfers are not initiated automatically
we need explicit nudge via this API call.

Testcases also became prone to timing as
the peer might be online but with libdrop off
which will render an immediate `CONN REFUSED` error.

Testcases are also tagged for easy running

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>