Improve scenario29-(10,11,12)
Libdrop doesn't have an event signalling
that the peer received the signal. We only have
RequestQueued which means the request materialized
with no errors and then we have TransferStarted which
indicates that the peer already called the norddrop_download()
API.

This leaves us hanging in case where we want to check
that the peer received the transfer without initiating any work
on it or if the peer went offline.

For this WaitForAnotherPeer() was modified
to accept the state it's looking for and in so we can use
it to detect that peer went offline and then initiate some work
like API calls.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>