Attempt to fix missing `FileDownloaded` event on rapidly disconnecting receiver
The issue seems to stem from the fact that for the event to be emitted, the `Started` event was required to be fired beforehand.
For the sender, the issue was not observed because there was no such requirement.
To sum up there should be no such requirement for the receiver and it's just a mistake on our side to impose it in the first place.

Signed-off-by: Mateusz Szczygieł <mateusz.szczygiel@nordsec.com>