Add checksum_events_granularity config
The problem consists in the fact that while transferring files
libdrop will generate ChecksumProgress events for every chunk of data
received (256KB) which can be a lot if the file is large.

Setting the checksum_events_granularity config makes the event
generation to occur at every checksum_events_granularity bytes of data
processed.

Signed-off-by: Daniel Fetti <daniel.fetti@nordsec.com>