Fix checksum progress for small granularity
There was a problem in the previous version if the checksum progress
granularity is set to a value smaller than 256KB. The progress bytes
were not reported correctly. This commit fixes it.

Also make the last checksum progress event to report the file size.

And also add more test scenarios.

Signed-off-by: Daniel Fetti <daniel.fetti@nordsec.com>