Call keystore.privkey() everytime we need it
Previously it was cached after the first call
and it so happens that the key might
sometimes change. Due to FFI layer issues, Libdrop
instance cannot be destroyed and recreated
thus leaving the only solution either a new API
or the easier one - passing KeyStore in Arc<>
and calling when needed.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>