Remove the migration for incoming_path_pending_states
macOS 11.5 has prebundled sqlite3 - 3.32.3 which we
need to support.

There are two features we use in 003 and 004 migrations that
is not supported there:

1. Use of `UPDATE-FROM` syntax which
is only supported since SQLite 2020-08-14 (3.33.0).

2. Use of ALTER TABLE DROP COLUMN which is
only supported since SQLite 2021-03-12 (3.35.0).

The 003 migration was introduced in v5.0.1 and 004 was
introduced in v5.1.0 both are used anymore by any platform
(earliest used v6.3.0 at the moment of writing). Meaning
the effort to migrate properly the UPDATE FROM is not worth it.

Signed-off-by: Lukas Pukenis <lukas.pukenis@nordsec.com>